Revert "Fix Pool Royale spin controller mapping"
